src/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Bcrypt
Flask-CORS
PyJWT
python-dotenv
PyYAML
reportlab
pandas
openpyxl
gunicorn
